time: today: collapse to simple "date -d today" and rm others
we were figuring out the exact second and then converting that to
today/yesterday/today but this is unnecessary becaues "date" itself
already knows these words.  furthermore, we want to add these as
actual shell commands in https://smemsh.net/src/pomoutils so we
remove "tomorrow" and "yesterday" because, after auditing, found
that we do not actually use these anyways (and "date -d tomorrow"
or "date -d yesterday" is simple enough inline).